[refactor] Contact 페이지 컴포넌트 폴더 이동
- 기존 pages와 routes 폴더가 혼돌될 수 있으므로 components의 pages폴더로 이동
