+++ front/src/pages/contact/index.ts
@@ -1,1 +1,1 @@
-export { default } from '@routes/Contact';
+export { default } from '@pages/Contact';
